requirements.txt
streamlit
folium
streamlit-folium
networkx
pandas
Pillow
ultralytics